Pushed v1.22 code
- Fixed crash when loading XMs with pattern lengths above 256. They are now
  safely truncated, and the user gets a warning message about the truncation.
- Allow loading overflown BPM/speed values from quirky XMs, up to BPM 999
  and speed 99 ("jk_error_txt.xm" is an example for quirky BPM/speed values).
- Tiny code cleanup

+++ src/ft2_audio.c
@@ -921,7 +921,7 @@ static void SDLCALL mixCallback(void *userdata, Uint8 *stream, int len)
 				pattSyncData.patternPos = song.curReplayerPattPos;
 				pattSyncData.pattern = song.curReplayerPattNr;
 				pattSyncData.songPos = song.curReplayerSongPos;
-				pattSyncData.speed = (uint8_t)song.speed;
+				pattSyncData.speed = song.speed;
 				pattSyncData.tempo = (uint8_t)song.tempo;
 				pattSyncData.globalVol = (uint8_t)song.globVol;
 				pattSyncData.timestamp = audio.tickTime64;
